Fix more import problems.
