+++ src/plugins/drawing/components/drawing-toolbar-buttons.tsx
@@ -6,8 +6,6 @@ import { ToolbarSettings } from "../model/drawing-basic-types";
 import SelectToolIcon from "../../../clue/assets/icons/select-tool.svg";
 import ColorFillIcon from "../assets/color-fill-icon.svg";
 import ColorStrokeIcon from "../assets/color-stroke-icon.svg";
-import GroupObjectsIcon from "../assets/group-objects-icon.svg";
-import UngroupObjectsIcon from "../assets/ungroup-objects-icon.svg";
 import DuplicateIcon from "../assets/duplicate-icon.svg";
 import DeleteSelectionIcon from "../../../assets/icons/delete/delete-selection-icon.svg";
 import { useTooltipOptions } from "../../../hooks/use-tooltip-options";
@@ -16,7 +14,6 @@ import { computeStrokeDashArray, IToolbarButtonProps, IToolbarManager,
   ToolbarModalButton } from "../objects/drawing-object";
 import { useTouchHold } from "../../../hooks/use-touch-hold";
 import SmallCornerTriangle from "../../../assets/icons/small-corner-triangle.svg";
-import { isGroupObject } from "../objects/group";
 
 interface IButtonClasses {
   modalButton?: ToolbarModalButton;
@@ -123,36 +120,10 @@ export const StrokeColorButton: React.FC<IColorButtonProps> = ({ settings, onCli
             settings={{ fill: settings.stroke, stroke }} onClick={onClick} />;
 };
 
-interface IActionButtonProps {
+export interface IActionButtonProps {
   toolbarManager: IToolbarManager;
 }
 
-export const GroupObjectsButton: React.FC<IActionButtonProps> = observer(function GroupObjectsButton({
-  toolbarManager 
-}) {
-  const onClick = () => {
-    toolbarManager.createGroup(toolbarManager.selection);
-  };
-  // Require at least two objects to group.
-  const disabled = toolbarManager.selection.length <= 1;
-  return <SvgToolbarButton SvgIcon={GroupObjectsIcon}  buttonClass="group" title="Group"
-          onClick={onClick} disabled={disabled} />;
-});
-
-export const UngroupObjectsButton: React.FC<IActionButtonProps> = observer(function UngroupObjectsButton({
-  toolbarManager
-}) {
-  const onClick = () => {
-    toolbarManager.ungroupGroups(toolbarManager.selection);
-  };
-  // Enabled if at least one selected object is a group
-  const disabled = !toolbarManager.selection.some((id)=> {
-    const selectedObj = toolbarManager.objectMap[id];
-    return selectedObj ? isGroupObject(selectedObj) : false;
-  });
-  return <SvgToolbarButton SvgIcon={UngroupObjectsIcon}  buttonClass="ungroup" title="Ungroup"
-          onClick={onClick} disabled={disabled} />;
-});
 
 export const DuplicateButton: React.FC<IActionButtonProps> = observer(function DuplicateButton({
   toolbarManager
